move plot to layouts
